move types to a better path
